// frontend-release/src/panels/lovelace/cards/types.ts
export interface AlarmPanelCardConfig extends LovelaceCardConfig {
  entity: string;
  name?: string;
  states?: string[];
  theme?: string;
}